move the right place
